Front End Completed... I'm horrible at commits...
The whole CRUD behaviour it's completed done in the front end now, it
uses Vuex has the main structure and has most of the functions that Vue
uses for CRUD to work.

There a little bug where, once a todo is updated, the index of todos in
the view is not updated, I'm still trying to figure out why.